% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lm.r
\name{lm2}
\alias{lm2}
\title{lm2}
\usage{
lm2(formula, data, subset, weights, na.action, method = "linmod",
  model = TRUE, x = FALSE, y = FALSE, qr = TRUE, singular.ok = TRUE,
  contrasts = NULL, offset, ..., check.rank = TRUE)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{subset}{}

\item{weights}{}

\item{na.action}{}

\item{method}{}

\item{model}{}

\item{x}{}

\item{y}{}

\item{qr}{}

\item{singular.ok}{}

\item{contrasts}{}

\item{offset}{}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{lm}.
}
\description{
Linear model fitter like \code{lm()}, but with the linmod package
driving the model fitting.
}

