% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rglm2-package.r
\docType{package}
\name{Rglm2-package}
\alias{Rglm2-package}
\title{Linear and Generalized Linear Model Fitters}
\description{
This package contains R's \code{lm()} and \code{glm()} functions modified
to use the linmod package to do the model fitting.
}
\details{
The primary purpose for not including these in the linmod package itself
is as a license barrier between the two packages.  Linmod is, and will
continue to be, more permissively licensed than the GPL.

\tabular{ll}{
   Package: \tab Rglm2 \cr
   Type: \tab Package \cr
   License: \tab GPL (>= 2) \cr
   NeedsCompilation: \tab no
}
}
\author{
Drew Schmidt
}
\references{
Project home page: \url{http://github.com/wrathematics/Rglm2}
}
\keyword{Package}

